rename parser.py -> lichen.py
