+++ clay_model/lichen.py
